+++ dataactvalidator/config/sqlrules/sqlRules.csv
@@ -163,5 +163,5 @@ FABS42.2,"PrimaryPlaceOfPerformanceForeignLocationDescription must be blank for 
 FABS43.1,"When PrimaryPlaceOfPerformanceCountryCode is not USA, PrimaryPlaceOfPerformanceCongressionalDistrict must be blank.",n,detached_award,fatal,,fabs43_detached_award_financial_assistance_1
 FABS43.2,"If PrimaryPlaceOfPerformanceZIP+4 is 5-digits or blank, a PrimaryPlaceOfPerformanceCongressionalDistrict must be provided.",n,detached_award,fatal,,fabs43_detached_award_financial_assistance_2
 FABS43.3,"If PrimaryPlaceOfPerformanceCongressionalDistrict is provided, it must be valid in the state or territory indicated by the PrimaryPlaceOfPerformanceCode. For PII-redacted non-aggregate records (i.e., when RecordType = 3), PrimaryPlaceOfPerformanceCongressionalDistrict must be blank.",n,detached_award,fatal,,fabs43_detached_award_financial_assistance_3
-FABS44.1,"If LegalEntityZIPLast4 is not provided and LegalEntityZIP5 is, LegalEntityCongressionalDistrict must be provided.",n,detached_award,fatal,,fabs44_detached_award_financial_assistance_1
-FABS44.2,"If LegalEntityCongressionalDistrict is provided, it must be valid in the 5-digit zip code indicated by LegalEntityZIP5.",n,detached_award,fatal,,fabs44_detached_award_financial_assistance_2
+FABS44.2,"If LegalEntityZIPLast4 is not provided and LegalEntityZIP5 is, LegalEntityCongressionalDistrict must be provided.",n,detached_award,fatal,,fabs44_detached_award_financial_assistance_2
+FABS44.3,"If LegalEntityCongressionalDistrict is provided, it must be valid in the 5-digit zip code indicated by LegalEntityZIP5.",n,detached_award,fatal,,fabs44_detached_award_financial_assistance_3
